Create Processing.java
just the start for the console program (not connected to the GUI. This is mostly things that we already had but needed it to run Administrator.java, Manager.java, and Regular_User.java. Those four programs are run together and should have the .jar file in the same folder when run. Start up the program by running Processing.java.

--Taylor